Abstracted Footer components Increased performance
