--- src/components/ClockDisplay.jsx
@@ -1,9 +1,0 @@
-import Clock from 'react-live-clock';
-
-const ClockDisplay = () => {
-  return (
-    <Clock format={'h:mm:ssa'} ticking={true} />
-  );
-};
-
-export default ClockDisplay;
